FEAT/custom useCollection hook added
