+++ src/pages/products/Products.js
@@ -1,38 +1,12 @@
 import React, { useState, useEffect } from 'react'
-import { projectFirestore } from '../../firebase/config';
-import Product from './Product';
+import { projectFirestore } from '../../firebase/config'; // FIREBASE
+import Product from './Product'; // COMPONENTS
+import { useCollection } from '../../hooks/useCollection';
 
 const Products = () => {
     //STATE
-    const [products, setProducts] = useState(null)
-    const [error, setError] = useState(null)
-    const [isPending, setIsPending] = useState(false)
+    const { data: products, error, isPending } = useCollection('products');
 
-    useEffect(() => {
-        // reset our state pre-fetching
-        setIsPending(true)
-        setError(null)
-
-        const unsubscribe = projectFirestore.collection('products')
-            .onSnapshot((snapshot) => {
-                if(snapshot.empty){
-                    setError('No documents cto load')
-                    setIsPending(false)
-                } else {
-                    let results = [];
-                    snapshot.forEach((doc) => {
-                        results.push({ ...doc.data(), id : doc.id })
-                    })
-                    setProducts(results)
-                    setIsPending(false)
-                }
-            }, (err) => {
-                setError(err.message)
-                console.log('Could not fetch data on this occasion')
-            })
-
-            return () => unsubscribe()
-    }, [])
 
     return(
         <section className='products-container'>
